+++ src/core/distributions/phylogenetics/tree/birthdeath/ConstantRateBirthDeathProcess.h
@@ -31,8 +31,8 @@ namespace RevBayesCore {
         
     public:
         ConstantRateBirthDeathProcess(const TypedDagNode<double> *o, const TypedDagNode<double> *s, const TypedDagNode<double> *e, 
-                                  const TypedDagNode<double> *r, const std::string& ss, const std::string &cdt, 
-                                  unsigned int nTaxa, const std::vector<std::string> &tn, const std::vector<Clade> &c);        
+                                      const TypedDagNode<double> *r, const std::string& ss, const std::string &cdt,
+                                      size_t nTaxa, const std::vector<std::string> &tn, const std::vector<Clade> &c);
         
         // public member functions
         ConstantRateBirthDeathProcess*                      clone(void) const;                                                                                  //!< Create an independent clone
