+++ src/core/monitors/admixture/AdmixtureBipartitionMonitor.h
@@ -62,14 +62,14 @@ namespace RevBayesCore {
         
     public:
         // Constructors and Destructors
-        AdmixtureBipartitionMonitor(TypedDagNode<AdmixtureTree> *t, TypedDagNode<double>* dr, TypedDagNode< std::vector< double > >* br, int ntr, int nar, int g, const std::string &fname, const std::string &del, bool pp=true, bool l=true, bool pr=true, bool ap = false, bool ci=false, bool ch=false);
+        AdmixtureBipartitionMonitor(TypedDagNode<AdmixtureTree> *t, TypedDagNode<double>* dr, TypedDagNode< std::vector< double > >* br, size_t ntr, size_t nar, unsigned long g, const std::string &fname, const std::string &del, bool pp=true, bool l=true, bool pr=true, bool ap = false, bool ci=false, bool ch=false);
         AdmixtureBipartitionMonitor(const AdmixtureBipartitionMonitor& f);
         
         // basic methods
         AdmixtureBipartitionMonitor*        clone(void) const;                                                  //!< Clone the object
         
         // Monitor functions
-        void                                monitor(long gen);                                                  //!< Monitor at generation gen
+        void                                monitor(unsigned long gen);                                         //!< Monitor at generation gen
         void                                swapNode(DagNode *oldN, DagNode *newN);
         
         // FileMonitor functions
