+++ src/core/moves/simple/SwitchRateJumpMove.cpp
@@ -88,7 +88,7 @@ double SwitchRateJumpMove::performSimpleMove( void )
     
     std::vector<double>& v = variable->getValue();
     // choose an index (do not propose a jump for the root)
-    index = int(rng->uniform01() * v.size());
+    index = size_t( rng->uniform01() * v.size() );
 
     // copy value
     storedValue = v[index];
