+++ src/core/moves/trees/compound/NearestNeighborInterchange.cpp
@@ -49,7 +49,7 @@ double NearestNeighborInterchange::performSimpleMove( void ) {
     TopologyNode* node;
     do {
         double u = rng->uniform01();
-        size_t index = std::floor(tau.getNumberOfNodes() * u);
+        size_t index = size_t( std::floor(tau.getNumberOfNodes() * u) );
         node = &tau.getNode(index);
     } while ( node->isRoot() || node->getParent().isRoot() );
