+++ src/revlanguage/datatypes/container/RealMatrix.cpp
@@ -87,7 +87,7 @@ bool RealMatrix::isConvertibleTo(const TypeSpec& type) const {
 /** Print value for user */
 void RealMatrix::printValue(std::ostream &o) const {
     
-    size_t previousPrecision = o.precision();
+    long previousPrecision = o.precision();
     std::ios_base::fmtflags previousFlags = o.flags();
     
     std::fixed( o );
